+++ src/Products/urban/migration/update_250.py
@@ -646,3 +646,14 @@ def set_page_style_for_mailing_templates(context):
             notify(ObjectModifiedEvent(template))
             logger.info("Set defaut page style for {}".format(template))
     logger.info("upgrade step done!")
+
+
+def install_browserlayer(context):
+    """ """
+    logger = logging.getLogger("urban: Install browserlayer")
+    logger.info("starting upgrade steps")
+    setup_tool = api.portal.get_tool("portal_setup")
+    setup_tool.runImportStepFromProfile(
+        "profile-Products.urban:default", "browserlayer"
+    )
+    logger.info("upgrade step done!")